Make rake db:sample_data more sensible, generate realistic-looking allele names

+++ lib/sample_data.rb
@@ -1,26 +1,29 @@
 class SampleData
   def self.load
+    raise "Sample data loading not supported in #{Rails.env} environment!" if ['staging', 'production'].include?(Rails.env)
+
     require 'factory_girl_rails'
 
     Centre.find_or_create_by_name('WTSI')
     Centre.find_or_create_by_name('ICS')
 
-    clone1 = Clone.find_by_clone_name('EPD_SAMPLE_1')
-    if ! clone1
-      clone1 = Factory.create(:clone, :pipeline_id => 1, :clone_name => 'EPD_SAMPLE_1')
-    end
+    [
+      {:clone_name => 'EPD_SAMPLE_1', :pipeline_id => 1},
+      {:clone_name => 'EPD_SAMPLE_2', :pipeline_id => 2}
+    ].each do |clone_data|
+      clone = Clone.find_by_clone_name(clone_data[:clone_name])
+      if clone
+        clone.mi_attempts.destroy_all
+        clone.destroy
+      end
 
-    clone2 = Clone.find_by_clone_name('EPD_SAMPLE_2')
-    if ! clone2
-      clone2 = Factory.create(:clone, :pipeline_id => 1, :clone_name => 'EPD_SAMPLE_2')
-    end
+      clone = Factory.create(:populated_clone, :pipeline_id => clone_data[:pipeline_id], :clone_name => clone_data[:clone_name])
 
-    [clone1, clone2].each do |clone|
-      clone.mi_attempts.destroy_all
       3.times do
-        Factory.create(:fully_populated_mi_attempt, :clone => clone)
+        Factory.create(:populated_mi_attempt, :clone => clone)
       end
     end
+
   end
 
 end
